# Conflicts:
#	jsapp/js/components/permissions/permParser.es6
#	jsapp/js/components/permissions/userAssetPermsEditor.es6

+++ kpi/views/v2/data.py
@@ -2,7 +2,12 @@
 from django.conf import settings
 from django.http import Http404
 from django.utils.translation import ugettext_lazy as _
-from rest_framework import renderers, serializers, viewsets
+from rest_framework import (
+    renderers,
+    serializers,
+    status,
+    viewsets,
+)
 from rest_framework.decorators import action
 from rest_framework.pagination import _positive_int as positive_int
 from rest_framework.response import Response
@@ -13,6 +18,7 @@ from kpi.exceptions import ObjectDeploymentDoesNotExist
 from kpi.models import Asset
 from kpi.paginators import DataPagination
 from kpi.permissions import (
+    DuplicateSubmissionPermission,
     EditSubmissionPermission,
     SubmissionPermission,
     SubmissionValidationStatusPermission,
@@ -150,6 +156,18 @@ class DataViewSet(AssetNestedObjectViewsetMixin, NestedViewSetMixin,
     >       curl -X GET https://[kpi]/api/v2/assets/aSAvYreNzVEkrWg5Gdcvg/data/234/edit/?return_url=false
 
 
+    ### Duplicate submission
+
+    Duplicates the data of a submission
+    <pre class="prettyprint">
+    <b>POST</b> /api/v2/assets/<code>{uid}</code>/data/<code>{id}</code>/duplicate/
+    </pre>
+
+    > Example
+    >
+    >       curl -X POST https://[kpi]/api/v2/assets/aSAvYreNzVEkrWg5Gdcvg/data/234/duplicate/
+
+
     ### Validation statuses
 
     Retrieves the validation status of a submission.
@@ -199,6 +217,45 @@ class DataViewSet(AssetNestedObjectViewsetMixin, NestedViewSetMixin,
     >        }
 
 
+    ### Bulk updating of submissions
+
+    <pre class="prettyprint">
+    <b>PATCH</b> /api/v2/assets/<code>{uid}</code>/data/bulk/
+    </pre>
+
+    > Example
+    >
+    >       curl -X PATCH https://[kpi]/api/v2/assets/aSAvYreNzVEkrWg5Gdcvg/data/bulk/
+
+    > **Payload**
+    >
+    >        {
+    >           "submission_ids": [{integer}],
+    >           "data": {
+    >               <field_to_update_1>: <value_1>,
+    >               <field_to_update_2>: <value_2>,
+    >               <field_to_update_n>: <value_n>
+    >           }
+    >        }
+
+    where `<field_to_update_n>` is a string and should be an existing XML field value of the submissions.
+    If `<field_to_update_n>` is part of a group or nested group, the field must follow the group hierarchy
+    structure, i.e.:
+
+    If the field is within a group called `group_1`, the field name is `question_1` and the new value is `new value`,
+    the payload should contain an item with the following structure:
+
+    <pre class="prettyprint">
+    "group_1/question_1": "new value"
+    </pre>
+
+    Similarly, if there are `N` nested groups, the structure will be:
+
+    <pre class="prettyprint">
+    "group_1/sub_group_1/.../sub_group_n/question_1": "new value"
+    </pre>
+
+
     ### CURRENT ENDPOINT
     """
 
@@ -221,12 +278,17 @@ class DataViewSet(AssetNestedObjectViewsetMixin, NestedViewSetMixin,
 
         return self.asset.deployment
 
-    @action(detail=False, methods=['DELETE'],
+    @action(detail=False, methods=['PATCH', 'DELETE'],
             renderer_classes=[renderers.JSONRenderer])
     def bulk(self, request, *args, **kwargs):
         deployment = self._get_deployment()
-        json_response = deployment.delete_submissions(request.data,
-                                                      request.user)
+        if request.method == 'DELETE':
+            json_response = deployment.delete_submissions(request.data,
+                                                          request.user)
+        elif request.method == 'PATCH':
+            json_response = deployment.bulk_update_submissions(
+                dict(request.data), request.user.id
+            )
         return Response(**json_response)
 
     def destroy(self, request, *args, **kwargs):
@@ -295,6 +357,19 @@ class DataViewSet(AssetNestedObjectViewsetMixin, NestedViewSetMixin,
             if not submission:
                 raise Http404
         return Response(submission)
+
+    @action(detail=True, methods=['POST'],
+            renderer_classes=[renderers.JSONRenderer],
+            permission_classes=[DuplicateSubmissionPermission])
+    def duplicate(self, request, pk, *args, **kwargs):
+        """
+        Creates a duplicate of the submission with a given `pk`
+        """
+        deployment = self._get_deployment()
+        duplicate_response = deployment.duplicate_submission(
+            requesting_user_id=request.user.id, instance_id=positive_int(pk)
+        )
+        return Response(duplicate_response, status=status.HTTP_201_CREATED)
 
     @action(detail=True, methods=['GET', 'PATCH', 'DELETE'],
             renderer_classes=[renderers.JSONRenderer],
